requirements.txt
flask>=2.0.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0  # for PDF generation
ollama>=0.1.0 